adding clink/_UartGeneric.py
